equipment model file created
